% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\name{get_education_data}
\alias{get_education_data}
\title{Obtain data from the Urban Institute Education Data Portal API}
\usage{
get_education_data(level = NULL, source = NULL, topic = NULL, ...,
  by = NULL, filters = NULL)
}
\arguments{
\item{level}{API data level to query}

\item{source}{API data source to query}

\item{topic}{API data topic to query}

\item{...}{Additional required parameters to pass to an API call}

\item{by}{Additional optional paramters to pass to an API call}

\item{filters}{Optional query values to filter an API call}
}
\value{
A `data.frame` of education data
}
\description{
Obtain data from the Urban Institute Education Data Portal API
}
\examples{
\dontrun{
library(educationdata)
df <- get_education_data(level = 'school-districts', source = 'ccd', topic = 'finance', year = 2004)
}

}
