% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_education_data_summary.R
\name{get_education_data_summary}
\alias{get_education_data_summary}
\title{Obtain data from the Urban Institute Education Data Portal API summary
endpoint functionality.}
\usage{
get_education_data_summary(
  level,
  source,
  topic = NULL,
  subtopic = NULL,
  stat = NULL,
  var = NULL,
  by = NULL,
  filters = NULL
)
}
\arguments{
\item{level}{API data level to query}

\item{source}{API data source to query}

\item{topic}{API data topic to query}

\item{subtopic}{Optional additional parameters to pass to an API call. Only
applicable to certain endpoints.}

\item{stat}{Summary statistic to be calculated. Valid options}

\item{var}{Variable to be summarized.}

\item{by}{Variable(s) to group results by.}

\item{filters}{Optional 'list' of query values to filter an API call}
}
\value{
A `data.frame` of aggregated education data
}
\description{
Obtain data from the Urban Institute Education Data Portal API summary
endpoint functionality.
}
