% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_education_data.R
\name{get_education_data}
\alias{get_education_data}
\title{Obtain data from the Urban Institute Education Data Portal API}
\usage{
get_education_data(
  level = NULL,
  source = NULL,
  topic = NULL,
  subtopic = NULL,
  by = NULL,
  filters = NULL,
  add_labels = FALSE,
  csv = FALSE
)
}
\arguments{
\item{level}{API data level to query}

\item{source}{API data source to query}

\item{topic}{API data topic to query}

\item{subtopic}{Optional 'list' of grouping parameters to pass to an API call}

\item{by}{DEPRECATED in favor of `subtopic`}

\item{filters}{Optional 'list' of query values to filter an API call}

\item{add_labels}{Add variable labels (when applicable)? Defaults to FALSE.}

\item{csv}{Download the full csv file? Defaults to FALSE.}
}
\value{
A `data.frame` of education data
}
\description{
Obtain data from the Urban Institute Education Data Portal API
}
